+++ src/DGtal/helpers/ShortcutsGeometry.h
@@ -48,6 +48,11 @@
 #include "DGtal/geometry/surfaces/estimation/IIGeometricFunctors.h"
 #include "DGtal/geometry/surfaces/estimation/IntegralInvariantVolumeEstimator.h"
 #include "DGtal/geometry/surfaces/estimation/IntegralInvariantCovarianceEstimator.h"
+
+#if defined(WITH_EIGEN)
+#include "DGtal/dec/DiscreteExteriorCalculusFactory.h"
+#include "DGtal/dec/ATSolver2D.h"
+#endif // defined(WITH_EIGEN)
 //////////////////////////////////////////////////////////////////////////////
 
 namespace DGtal
@@ -175,14 +180,17 @@ namespace DGtal
       // ----------------------- Static services --------------------------------------
     public:
 
-      // ----------------------- True geometry services ------------------------------
+      // ----------------------- Exact geometry services ------------------------------
+      /// @name Exact geometry services
+      /// @{
     public:
 
       /// @return the parameters used in ShortcutsGeometry.
       static Parameters defaultParameters()
       {
         return parametersShapeGeometry()
-          | parametersGeometryEstimation();
+          | parametersGeometryEstimation()
+          | parametersATApproximation();
       }
     
       /// @return the parameters and their default values which are used
@@ -386,9 +394,11 @@ namespace DGtal
         return n_true_estimations;
       }
     
-    
+      /// @}
     
       // --------------------------- geometry estimation ------------------------------
+      /// @name Geometry estimation services
+      /// @{
     public:
 
       /// @return the parameters and their default values which are used
@@ -965,11 +975,301 @@ namespace DGtal
                              std::back_inserter( mc_estimations ) );
           return mc_estimations;
         }
-    
 
-    
+      /// @}
+
+      // --------------------------- AT approximation ------------------------------
+      /// @name AT approximation services
+      /// @{
+    public:
+
+      /// @return the parameters and their default values which are used
+      /// to compute Ambrosio-Tortorelli piecewise-smooth approximation of a function.
+      ///   - at-enabled      [  1     ]: 1 if AT is enabled (WITH_EIGEN), 0 otherwise.
+      ///   - at-alpha        [  0.1   ]: parameter alpha in AT (data fit)
+      ///   - at-lambda       [  0.025 ]: parameter lambda in AT (1/length of discontinuities)
+      ///   - at-epsilon      [  0.5   ]: (last value of) parameter epsilon in AT (width of discontinuities)
+      ///   - at-epsilon-start[  2.0   ]: first value for parameter epsilon in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-epsilon-ratio[  2.0   ]: ratio between two consecutive epsilon value in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-max-iter     [ 10     ]: maximum number of alternate minization in AT optimization
+      ///   - at-diff-v-max   [  0.0001]: stopping criterion that measures the loo-norm of the evolution of \a v between two iterations
+      ///   - at-v-policy     ["Maximum"]: the policy when outputing feature vector v onto cells: "Average"|"Minimum"|"Maximum"
+      ///
+      /// @note Requires Eigen linear algebra backend. `Use cmake -DWITH_EIGEN=true ..`
+      static Parameters parametersATApproximation()
+      {
+#if defined(WITH_EIGEN)
+        return Parameters
+          ( "at-enabled",        1 )
+          ( "at-alpha",          0.1 )
+          ( "at-lambda",         0.025 )
+          ( "at-epsilon",        0.25 )
+          ( "at-epsilon-start",  2.0 )
+          ( "at-epsilon-ratio",  2.0 )
+          ( "at-max-iter",      10 )
+          ( "at-diff-v-max",     0.0001 )
+          ( "at-v-policy",   "Maximum" );
+#else // defined(WITH_EIGEN)
+        return Parameters( "at-enabled", 0 );
+#endif// defined(WITH_EIGEN)
+      }
+
+#if defined(WITH_EIGEN)
+
+      /// Given any digital \a surface, a surfel range \a surfels, and an input vector field \a input,
+      /// returns a piece-smooth approximation of \a input using Ambrosio-Tortorelli functional.
+      ///
+      /// @see \ref moduleAT
+      ///
+      /// @tparam TAnyDigitalSurface either kind of DigitalSurface, like ShortcutsGeometry::LightDigitalSurface or ShortcutsGeometry::DigitalSurface.
+      /// @tparam VectorFieldInput the type of vector field for input values (RandomAccess container)
+      ///
+      /// @param[in] surface the digital surface
+      /// @param[in] surfels the sequence of surfels at which we compute the normals
+      /// @param[in] params the parameters:
+      ///   - at-alpha        [  0.1   ]: parameter alpha in AT (data fit)
+      ///   - at-lambda       [  0.025 ]: parameter lambda in AT (1/length of discontinuities)
+      ///   - at-epsilon      [  0.5   ]: (last value of) parameter epsilon in AT (width of discontinuities)
+      ///   - at-epsilon-start[  2.0   ]: first value for parameter epsilon in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-epsilon-ratio[  2.0   ]: ratio between two consecutive epsilon value in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-max-iter     [ 10     ]: maximum number of alternate minization in AT optimization
+      ///   - at-diff-v-max   [  0.0001]: stopping criterion that measures the loo-norm of the evolution of \a v between two iterations
+      /// @param[in] input the input vector field (a vector of vector values)
+      ///
+      /// @return the piecewise-smooth approximation of \a input.
+      ///
+      /// @note Requires Eigen linear algebra backend. `Use cmake -DWITH_EIGEN=true ..`
+      template <typename TAnyDigitalSurface,
+                typename VectorFieldInput>
+      static
+      VectorFieldInput
+      getATVectorFieldApproximation( CountedPtr<TAnyDigitalSurface> surface,
+                                     const SurfelRange&             surfels,
+                                     const VectorFieldInput&        input,
+                                     const Parameters&              params
+                                     = parametersATApproximation() | parametersGeometryEstimation() )
+      {
+        int      verbose   = params[ "verbose"          ].as<int>();
+        Scalar   alpha_at  = params[ "at-alpha"         ].as<Scalar>();
+        Scalar   lambda_at = params[ "at-lambda"        ].as<Scalar>();
+        Scalar   epsilon1  = params[ "at-epsilon-start" ].as<Scalar>();
+        Scalar   epsilon2  = params[ "at-epsilon"       ].as<Scalar>();
+        Scalar   epsilonr  = params[ "at-epsilon-ratio" ].as<Scalar>();
+        int      max_iter  = params[ "at-max-iter"      ].as<int>();
+        Scalar   diff_v_max= params[ "at-diff-v-max"    ].as<Scalar>();
+        typedef DiscreteExteriorCalculusFactory<EigenLinearAlgebraBackend> CalculusFactory;
+        const auto calculus = CalculusFactory::createFromNSCells<2>( surfels.cbegin(), surfels.cend() );
+        ATSolver2D< KSpace > at_solver( calculus, verbose );
+        at_solver.initInputVectorFieldU2( input, surfels.cbegin(), surfels.cend() );
+        at_solver.setUp( alpha_at, lambda_at );
+        at_solver.solveGammaConvergence( epsilon1, epsilon2, epsilonr, false, diff_v_max, max_iter );
+        auto output = input;
+        at_solver.getOutputVectorFieldU2( output, surfels.cbegin(), surfels.cend() );
+        return output;
+      }
+
+      /// Given any digital \a surface, a surfel range \a surfels, and an input vector field \a input,
+      /// returns a piece-smooth approximation of \a input using Ambrosio-Tortorelli functional.
+      /// Given a range of pointels, linels or 2-cells [itB,itE), it
+      /// also outputs the feature vector \a features, corresponding to
+      /// 0-form \a v in AT (the average of \a v for linels/surfels).
+      ///
+      /// @see \ref moduleAT
+      ///
+      /// @tparam TAnyDigitalSurface either kind of DigitalSurface, like ShortcutsGeometry::LightDigitalSurface or ShortcutsGeometry::DigitalSurface.
+      /// @tparam VectorFieldInput the type of vector field for input values (RandomAccess container)
+      /// @tparam CellRangeConstIterator the type of iterator for traversing a range of cells
+      ///
+      /// @param[out] features the vector of scalar feature values (a scalar field where 1 means continuity and 0 discontinuity in the reconstruction), evaluated in the range[itB,itE).
+      /// @param[in] itB the start of the range of cells.
+      /// @param[in] itE past the end of the range of cells.
+      /// @param[in] surface the digital surface
+      /// @param[in] surfels the sequence of surfels at which we compute the normals
+      /// @param[in] params the parameters:
+      ///   - at-alpha        [  0.1   ]: parameter alpha in AT (data fit)
+      ///   - at-lambda       [  0.025 ]: parameter lambda in AT (1/length of discontinuities)
+      ///   - at-epsilon      [  0.5   ]: (last value of) parameter epsilon in AT (width of discontinuities)
+      ///   - at-epsilon-start[  2.0   ]: first value for parameter epsilon in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-epsilon-ratio[  2.0   ]: ratio between two consecutive epsilon value in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-max-iter     [ 10     ]: maximum number of alternate minization in AT optimization
+      ///   - at-diff-v-max   [  0.0001]: stopping criterion that measures the loo-norm of the evolution of \a v between two iterations
+      ///   - at-v-policy     ["Maximum"]: the policy when outputing feature vector v onto cells: "Average"|"Minimum"|"Maximum"
+      /// @param[in] input the input vector field (a vector of vector values)
+      ///
+      /// @return the piecewise-smooth approximation of \a input.
+      ///
+      /// @note Requires Eigen linear algebra backend. `Use cmake -DWITH_EIGEN=true ..`
+      template <typename TAnyDigitalSurface,
+                typename VectorFieldInput,
+                typename CellRangeConstIterator>
+      static
+      VectorFieldInput
+      getATVectorFieldApproximation( Scalars&                       features,
+                                     CellRangeConstIterator         itB,
+                                     CellRangeConstIterator         itE,
+                                     CountedPtr<TAnyDigitalSurface> surface,
+                                     const SurfelRange&             surfels,
+                                     const VectorFieldInput&        input,
+                                     const Parameters&              params
+                                     = parametersATApproximation() | parametersGeometryEstimation() )
+      {
+        int      verbose   = params[ "verbose"          ].as<int>();
+        Scalar   alpha_at  = params[ "at-alpha"         ].as<Scalar>();
+        Scalar   lambda_at = params[ "at-lambda"        ].as<Scalar>();
+        Scalar   epsilon1  = params[ "at-epsilon-start" ].as<Scalar>();
+        Scalar   epsilon2  = params[ "at-epsilon"       ].as<Scalar>();
+        Scalar   epsilonr  = params[ "at-epsilon-ratio" ].as<Scalar>();
+        int      max_iter  = params[ "at-max-iter"      ].as<int>();
+        Scalar   diff_v_max= params[ "at-diff-v-max"    ].as<Scalar>();
+        std::string policy = params[ "at-v-policy"      ].as<std::string>();
+        typedef DiscreteExteriorCalculusFactory<EigenLinearAlgebraBackend> CalculusFactory;
+        const auto calculus = CalculusFactory::createFromNSCells<2>( surfels.cbegin(), surfels.cend() );
+        ATSolver2D< KSpace > at_solver( calculus, verbose );
+        at_solver.initInputVectorFieldU2( input, surfels.cbegin(), surfels.cend() );
+        at_solver.setUp( alpha_at, lambda_at );
+        at_solver.solveGammaConvergence( epsilon1, epsilon2, epsilonr, false, diff_v_max, max_iter );
+        auto output = input;
+        at_solver.getOutputVectorFieldU2( output, surfels.cbegin(), surfels.cend() );
+        auto p = ( policy == "Average" ) ? at_solver.Average
+          :      ( policy == "Minimum" ) ? at_solver.Minimum
+          :                                at_solver.Maximum;
+        at_solver.getOutputScalarFieldV0( features, itB, itE, p );
+        return output;
+      }
+
+      /// Given any digital \a surface, a surfel range \a surfels, and
+      /// an input scalar field \a input, returns a piece-smooth
+      /// approximation of \a input using Ambrosio-Tortorelli
+      /// functional.
+      ///
+      /// @see \ref moduleAT
+      ///
+      /// @tparam TAnyDigitalSurface either kind of DigitalSurface, like ShortcutsGeometry::LightDigitalSurface or ShortcutsGeometry::DigitalSurface.
+      ///
+      /// @param[in] surface the digital surface
+      /// @param[in] surfels the sequence of surfels at which we compute the normals
+      /// @param[in] params the parameters:
+      ///   - at-alpha        [  0.1   ]: parameter alpha in AT (data fit)
+      ///   - at-lambda       [  0.025 ]: parameter lambda in AT (1/length of discontinuities)
+      ///   - at-epsilon      [  0.5   ]: (last value of) parameter epsilon in AT (width of discontinuities)
+      ///   - at-epsilon-start[  2.0   ]: first value for parameter epsilon in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-epsilon-ratio[  2.0   ]: ratio between two consecutive epsilon value in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-max-iter     [ 10     ]: maximum number of alternate minization in AT optimization
+      ///   - at-diff-v-max   [  0.0001]: stopping criterion that measures the loo-norm of the evolution of \a v between two iterations
+      /// @param[in] input the input scalar field (a vector of scalar values)
+      ///
+      /// @return the piecewise-smooth approximation of \a input.
+      ///
+      /// @note Requires Eigen linear algebra backend. `Use cmake -DWITH_EIGEN=true ..`
+      template <typename TAnyDigitalSurface>
+      static
+      Scalars
+      getATScalarFieldApproximation( CountedPtr<TAnyDigitalSurface> surface,
+                                     const SurfelRange&             surfels,
+                                     const Scalars&                 input,
+                                     const Parameters&              params
+                                     = parametersATApproximation() | parametersGeometryEstimation() )
+      {
+        int      verbose   = params[ "verbose"          ].as<int>();
+        Scalar   alpha_at  = params[ "at-alpha"         ].as<Scalar>();
+        Scalar   lambda_at = params[ "at-lambda"        ].as<Scalar>();
+        Scalar   epsilon1  = params[ "at-epsilon-start" ].as<Scalar>();
+        Scalar   epsilon2  = params[ "at-epsilon"       ].as<Scalar>();
+        Scalar   epsilonr  = params[ "at-epsilon-ratio" ].as<Scalar>();
+        int      max_iter  = params[ "at-max-iter"      ].as<int>();
+        Scalar   diff_v_max= params[ "at-diff-v-max"    ].as<Scalar>();
+        typedef DiscreteExteriorCalculusFactory<EigenLinearAlgebraBackend> CalculusFactory;
+        const auto calculus = CalculusFactory::createFromNSCells<2>( surfels.cbegin(), surfels.cend() );
+        ATSolver2D< KSpace > at_solver( calculus, verbose );
+        at_solver.initInputScalarFieldU2( input, surfels.cbegin(), surfels.cend() );
+        at_solver.setUp( alpha_at, lambda_at );
+        at_solver.solveGammaConvergence( epsilon1, epsilon2, epsilonr, false, diff_v_max, max_iter );
+        auto output = input;
+        at_solver.getOutputScalarFieldU2( output, surfels.cbegin(), surfels.cend() );
+        return output;
+      }
+
+      /// Given any digital \a surface, a surfel range \a surfels, and
+      /// an input scalar field \a input, returns a piece-smooth
+      /// approximation of \a input using Ambrosio-Tortorelli
+      /// functional.  Given a range of pointels, linels or 2-cells
+      /// [itB,itE), it also outputs the feature vector \a features,
+      /// corresponding to 0-form \a v in AT (the average of \a v for
+      /// linels/surfels).
+      ///
+      /// @see \ref moduleAT
+      ///
+      /// @tparam TAnyDigitalSurface either kind of DigitalSurface, like ShortcutsGeometry::LightDigitalSurface or ShortcutsGeometry::DigitalSurface.
+      /// @tparam CellRangeConstIterator the type of iterator for traversing a range of cells
+      ///
+      /// @param[out] features the vector of scalar feature values (a
+      /// scalar field where 1 means continuity and 0 discontinuity in
+      /// the reconstruction), evaluated in the range[itB,itE).
+      ///
+      /// @param[in] itB the start of the range of cells.
+      /// @param[in] itE past the end of the range of cells.
+      /// @param[in] surface the digital surface
+      /// @param[in] surfels the sequence of surfels at which we compute the normals
+      /// @param[in] params the parameters:
+      ///   - at-alpha        [  0.1   ]: parameter alpha in AT (data fit)
+      ///   - at-lambda       [  0.025 ]: parameter lambda in AT (1/length of discontinuities)
+      ///   - at-epsilon      [  0.5   ]: (last value of) parameter epsilon in AT (width of discontinuities)
+      ///   - at-epsilon-start[  2.0   ]: first value for parameter epsilon in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-epsilon-ratio[  2.0   ]: ratio between two consecutive epsilon value in Gamma-convergence optimization (sequence of AT optimization with decreasing epsilon)
+      ///   - at-max-iter     [ 10     ]: maximum number of alternate minization in AT optimization
+      ///   - at-diff-v-max   [  0.0001]: stopping criterion that measures the loo-norm of the evolution of \a v between two iterations
+      ///   - at-v-policy     ["Maximum"]: the policy when outputing feature vector v onto cells: "Average"|"Minimum"|"Maximum"
+      /// @param[in] input the input scalar field (a vector of scalar values)
+      ///
+      /// @return the piecewise-smooth approximation of \a input.
+      ///
+      /// @note Requires Eigen linear algebra backend. `Use cmake -DWITH_EIGEN=true ..`
+      template <typename TAnyDigitalSurface,
+                typename CellRangeConstIterator>
+      static
+      Scalars
+      getATScalarFieldApproximation( Scalars&                       features,
+                                     CellRangeConstIterator         itB,
+                                     CellRangeConstIterator         itE,
+                                     CountedPtr<TAnyDigitalSurface> surface,
+                                     const SurfelRange&             surfels,
+                                     const Scalars&                 input,
+                                     const Parameters&              params
+                                     = parametersATApproximation() | parametersGeometryEstimation() )
+      {
+        int      verbose   = params[ "verbose"          ].as<int>();
+        Scalar   alpha_at  = params[ "at-alpha"         ].as<Scalar>();
+        Scalar   lambda_at = params[ "at-lambda"        ].as<Scalar>();
+        Scalar   epsilon1  = params[ "at-epsilon-start" ].as<Scalar>();
+        Scalar   epsilon2  = params[ "at-epsilon"       ].as<Scalar>();
+        Scalar   epsilonr  = params[ "at-epsilon-ratio" ].as<Scalar>();
+        int      max_iter  = params[ "at-max-iter"      ].as<int>();
+        Scalar   diff_v_max= params[ "at-diff-v-max"    ].as<Scalar>();
+        std::string policy = params[ "at-v-policy"      ].as<std::string>();
+        typedef DiscreteExteriorCalculusFactory<EigenLinearAlgebraBackend> CalculusFactory;
+        const auto calculus = CalculusFactory::createFromNSCells<2>( surfels.cbegin(), surfels.cend() );
+        ATSolver2D< KSpace > at_solver( calculus, verbose );
+        at_solver.initInputScalarFieldU2( input, surfels.cbegin(), surfels.cend() );
+        at_solver.setUp( alpha_at, lambda_at );
+        at_solver.solveGammaConvergence( epsilon1, epsilon2, epsilonr, false, diff_v_max, max_iter );
+        auto output = input;
+        at_solver.getOutputScalarFieldU2( output, surfels.cbegin(), surfels.cend() );
+        auto p = ( policy == "Average" ) ? at_solver.Average
+          :      ( policy == "Minimum" ) ? at_solver.Minimum
+          :                                at_solver.Maximum;
+        at_solver.getOutputScalarFieldV0( features, itB, itE, p );
+        return output;
+      }
+
+#endif // defined(WITH_EIGEN)
+      
+      /// @}
+      
       // ------------------------- Error measures services -------------------------
-
+      /// @name Error measure services
+      /// @{
+    public:
+      
       /// Orient \a v so that it points in the same direction as \a
       /// ref_v (scalar product is then non-negative afterwards).
       ///
@@ -1090,7 +1390,11 @@ namespace DGtal
         return loo;
       }
 
+      /// @}
+
       // ----------------------- Standard services ------------------------------
+      /// @name Standard services
+      /// @{
     public:
 
       /**
@@ -1129,6 +1433,8 @@ namespace DGtal
        */
       ShortcutsGeometry & operator= ( ShortcutsGeometry && other ) = delete;
 
+      /// @}
+      
       // ----------------------- Interface --------------------------------------
     public: